+++ src/mc/network/packet/CorrectPlayerMovePredictionPacket.h
@@ -13,7 +13,7 @@ class CorrectPlayerMovePredictionPacket : public ::Packet {
 public:
     Vec3               mPos;      // this+0x30
     Vec3               mPosDelta; // this+0x3C
-    unsigned long long mTick;     // this+0x48
+    uint64_t mTick;     // this+0x48
     bool               mOnGround; // this+0x50
 
     // prevent constructor by default
